+++ changelog.md
@@ -9,11 +9,13 @@
 - removed ability to backup old messages
 - fixed some typos
 - fixed a bug due to which if you passed an argument with the path for the config to a bat file, it was lost and not passed to the python script
+- fixed bug due to which files with the same name overwrite each other
+- fixed bug with duplicate messages
 - config become more detailed
 - added ability to backup multiple chats
-- added ability to just forward message instead of downloading it to the storage
+<!-- - added ability to just forward message instead of downloading it to the storage -->
 - added update checker
-- added ability to turn on auto updates
+- added ability to turn on automatic updates
 - most of the code has been rewritten, cleaned up and optimized
 
 ## 2.1